feat(parsz): migrate .integer/.float to .i64/.f64 and add narrow type integration tests
Update all test definitions and example to use new ValueType variants.
Add integration tests for u8, i8, u16, u32, i32, f32 types including
boundary values, defaults, multiple args, and OOM safety.

+++ examples/sample.zig
@@ -25,7 +25,7 @@ const cmd = parsz.Command{
         .{
             .name = "count",
             .kind = .option,
-            .value_type = .integer,
+            .value_type = .i64,
             .long = "count",
             .default = "1",
             .help = "Repeat count",
